+++ tests/testthat/test-ATEcomplete.R
@@ -3,5 +3,5 @@ test_that("read ATE data sets", {
   d1 <- ATE.complete(fileName, TRUE)
   expect_equal(nrow(d1),1)
   d2 <- ATE.complete(fileName)
-  expect_equal(nrow(d2),103)
+  expect_equal(nrow(d2),115)
 })