+++ tests/testthat/test-Status.R
@@ -2,5 +2,5 @@ test_that("Test Status File", {
   fileName <- ATE.sampleFiles('_Status')
   df <- ATE.status(fileName)
 
-  expect_equal(nrow(df), 19)
+  expect_equal(nrow(df), 171)
 })